requirements.txt
# Django核心
Django>=5.0,<5.2
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# 图像处理与AI
opencv-python>=4.9.0
numpy>=1.24.0
pillow>=10.0.0
tensorflow>=2.15.0
scikit-learn>=1.4.0

# 任务调度
django-celery-beat>=2.5.0
celery>=5.3.0
redis>=5.0.0

# 工具库
python-dotenv>=1.0.0
pytz>=2024.1
python-dateutil>=2.8.0